Force stop of page tree traversal when max level reached
The previous implementation, FindPDFPage, was already doing this since
the recursive call was always with return. Currently, we were trying to
keep going even after reaching max level. The problem is that if the
page tree is not a tree, we might loop forever. This could also be
solved by keeping track of the dictionaries that have been visited, but
this solution takes much less space.

BUG=672172

Change-Id: Ia37aea58e92b6068de69f26736c612aa6a0ff4b3