Remove the ::GetCapacity methods.
The GetCapacity methods return a void* because they return different types of
internal class memory based on what the calling parameter was. This is confusing
and makes it difficult to tell when then enum values can be removed.

This CL removes GetCapacity and adds methods as needed to get the real values.

Change-Id: I64c2edc858220624880e27f4ed49c2dae080f462