Verify value before doing bitshift
Given a crafted PDF it's possible to get a BITSHIFT operator where the shift
value is outside the valid range. Verify we will not shift too far before
doing the shift.

It's also possble the shift value will be larger then an int. We store the
shift into a float and then cast back to an int when we know it is small enough.

BUG=chromium:681091

Change-Id: I765b499196adccf1ea4fbfc1ccaaa411484f6245