Fix CSS fuzzer input size
Currently we use the size provided by clusterfuzz when initializing the
css syntax parser. This maybe incorrect as the CFX_WideString may have a
different count after converting to UTF. Use the wide string length instead
of the provided size.

We need to guard against strings that convert to blank when doing the wide
conversion so add an early exit.

BUG=682551

Change-Id: I3e014647fcf869681098a1b4446306b8b3eb9323