Fixup test harness
When the results of the test runner were converted to return a tuple the
following code which checked the results for true/false were not updated
to extract the success value from the tuple. This caused the code to think
that the results always passed even when they failed.

This CL fixes the two tests which were broken (both just minor image result
differences) which slipped in during this breakage.

Change-Id: I01b56dd7b05013c2c12c83543746cf59b145e561