Allow building XFA without additional codecs.
This is something we'd like to try for initial XFA launches
adding in codecs as justified by results in the wild.

Adding statistics for the unsupported cases is a follow-up
exercise once this builds correctly.

We always build all the additional libraries, to allow
fuzzers to link against them even if we are not shipping
them.  The linker will sort it out for the actual code.

That the existing tests seem to pass with the codecs
disabled warrants further investigation.

Change-Id: Iad269db91289f12dc9f5dda8f48121d27a0c4367