Update patch set 5
Patch Set 5:

EmbedderTest does not init a document in SetUp, and methods like OpenDocument assume that document_ is not initialized (if it is, it will leak). So I moved the call of FPDF_CreateNewDocument to a method instead.

Patch-set: 5
Reviewer: Nicolás Peña <6295@fca2d64a-0420-3532-b200-12848c95a47b>