Remove CFX_StackTemplate in xfa_utils.h.
Tree traversal doesn't require a stack as we always have
parent pointers in nodes.

Break the code into smaller units to help make more
comprehensible what is going on.

Remove init methods and force callers to pass us root
in the constructor.

The operations are complicated by the lack of a last child
or previous sibling pointer in our nodes.

Change-Id: Ia32bb44354352686cb79ccd67fe8aa5cd5ccef58