Move to using shared FreeType on Windows
Add a public config to fxfreetype to make includes visible and provide a
freetype_common group as a public dependency on fxcrt. freetype_common
switches between fxfreetype and //third_party/freetype, which then
originates from the external Chromium FreeType. Make using the external
FreeType the default configuration for Chromium on Windows as well.

BUG=chromium:700926

Change-Id: I73ae26979dcf69a419485def23c7a13dffa2a15d