Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can you explain what the differences were between the TXT and RTF types?

The difference in the Line classes was the use of CFX_TxtChar and CFX_TxtPiece vs CFX_RTFChar and CFX_RTFPiece. There were a few methods added, but the ones that existed worked exactly the same.

The difference in the Break classes is that there were a couple of members that existed in one but not the other and there were a few methods that needed to be added. (The two font ints, the char styles, dwIdentity along with the retained user_data pointer.)

The Char classes were the same, just stuff that existed in one or the other that got merged together.

Patch-set: 2
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>