Convert CPDF_Creator::m_Pos to uint32_t
Currently m_Pos is a FX_POSITION. The value stored into the void* is
typically a uint32_t. We sometimes read that uint32_t value as a size_t
to compare against vector lengths. This Cl converts all usages to be
uint32_t and uses pdfium::CollectionSize to get the correct typed vector
size.

Change-Id: I5f452817ab6049e03477945e4622c6911c89f877