Add array_buffer JS test.
JS Array Buffers are the first candidate to be allocated from
PartitionAlloc when it becomes available, so add test first.

Presently, we will return as large an array buffer as the system
can handle; this is generally a bad idea so limit them to 256MB
and test that we handle failure.

Change-Id: I205745a7938d69eb32ac883b90824f2f9e584ec7