Update patch set 1
Patch Set 1:

I will next fix the build for Mac, seems to be trivial use of implicit declaration of functions, as the code is K&R and it is being built as C99.

From the bot:
gzwrite.c:89:20: error: implicit declaration of function 'write' is invalid in C99 [-Werror,-Wimplicit-function-declaration]

We can either disable the compiler flag (e.g. -Wimplicit-function-declaration) for zlib *or* patch it to include the header with the definition.

I would rather avoid patching zlib but on the other hand the warning is justified.

Patch-set: 1
Reviewer: Adenilson Cavalcanti <6735@fca2d64a-0420-3532-b200-12848c95a47b>