fix new tab crash in skia clip stack
The crash on the new tab page is triggered by processing a transparency.

This creates a new Skia device in CPDF_RenderStatus::LoadSMask():

// cpdf_renderstatus.cpp # 2557
  if (!bitmap_device.Create(width, height, format, nullptr))

which sets the Skia clip stack to empty.

It then calls
   RenderObjectList()
      RenderSingleObject()
         ProcessClipPath()

which resets the clip stack;

// cpdf_renderstatus.cpp # 1882
  m_LastClipPath = ClipPath;
  m_pDevice->RestoreState(true);

At this point m_LastClipPath contains {m_Ref={m_pObject={m_pObj=empty } } }

The impelemntation in  CFX_AggDeviceDriver::RestoreState() is

// fx_agg_driver.cpp # 1283
  if (m_StateStack.empty())
    return;

This hides unbalanced save/restores, but reworking PDFium to balance is nontrivial.


R=dsinclair@chromium.org
BUG=chromium:704442

Bug:

Change-Id: Ia70d4dd7bd118e40adc5c029acbaa0b66372d3aa