Update patch set 1
Patch Set 1: Commit-Queue+1

PTAL. This is followup from the last CL to fixup the reference inputs and default values. There are a lot of nits in the files edited so probably best to focus on the code changed. We'll need to do a nit pass in the XFA code at some point.

Patch-set: 1
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1