Avoid one more instance of the anti-pattern in 706346.
Need to pass a const CFX_RetainPtr<X>& reference into a function,
assign a different object to a m_ variable of the same type,
continue to use the reference after the assignment, and call the
function elsewhere with the same m_ variable as an argument.

Unclear if it is called in this manner, but the first three points
hold, so be proactive.

Change-Id: I0ece4d7da0b8cf5f3079c53fa612ca07e9632502