Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Meh, like bushy tree.  Do we have other sub-dirs in fxcrt?  I don't think so.


No, but we have a whole bunch of xml/ stuff. There is other xml/ gunk still in XFA that i'm not sure if it can be split apart. The benefit I see of this is that we have the xml/ directory and we can see all the xml related bits. Hopefully we can find things to cleanup from there.

Patch-set: 1
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0