Use correct length in guard check
When fixing crbug.com/672177 we added a guard that we aren't reading off
the end of the file. That guard used the file access Position(). This is
the wrong value to compare against as our read position and the file
access Position maybe different. This CL updates the check to use the
correct current file position.

Bug: pdfium:697
Change-Id: I68a5eaed2f1f3d65422605f0a8474144cfa7d172