Update patch set 1
Patch Set 1: Commit-Queue+1

PTAL (there is no test as I haven't figured out how to make a tagged PDF yet without hand writing a .in file ...).

I'm wondering if this is the best way to do this or not. There are a few other strings in a Struct Element that would make sense to expose:
  - ID  (byte string)
  - Lang  (a language identifier)
  - E     (the expanded form of an abbreviation
  - ActualText  (text that is an exact replacement for the element and children).

We already have Alt and this CL adds T.

Along with those, there is also a /O which is the owners array/dict. Which is basically a set of dictionaries that will have various keys that I haven't figure out how to expose through the API yet. (This is the CPDF_StructElement::GetAttr and FindAttrDict code internally). The /O dictionaries are the ones with interesting things like /Width and /Height, /BBox, etc.

So, thoughts on landing this as is, or should I sit down and try to figure out the full set of APIs for tagged documents? The harm I see of landing this is that we've 'minted' the API, thought it will be very very new, so we could possibly change it soon-ish if we really needed too.

Patch-set: 1
Reviewer: Tom Sepez <5495@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Lei Zhang <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+1