Remove the "old way" of performing FlateUncompress.
Repeated looping over Realloc() with a linearly increasing buffer
size is going to be O(N^2) which is why it was limited to small
cases.  But it's wrong to do it anyways, and it kicks out a
lurking issue in PartitionAlloc as it turns out.

See https://crbug.com/709271, but this doesn't actually fix that
issue, it just avoids it.

Change-Id: I39d8f8df0d5fc358b7b9caa15c507a8f57e45004