Update patch set 1
Patch Set 1:

Legal info: this is the first pdfium patch I created in my company time. Collabora (Productivity) submitted a CLA as far as I know, created the "collaboraproductivity" Google Group and the email I used for this patch is part of that group. I hope that's enough to address requirements.

Other info: when LibreOffice bundles pdfium we have a similar patch:

https://cgit.freedesktop.org/libreoffice/core/tree/external/pdfium/system-zlib.patch.1

This change is an attempt to upstream this.

Patch-set: 1
Reviewer: Miklos V <6835@fca2d64a-0420-3532-b200-12848c95a47b>