Subclass the stream implementations from CFGAS_Stream
This CL renames CFGAS_TextStream to CFGAS_Stream. Then, CFGAS_Stream,
instead of accepting an IFGAS_StreamImp is converted to a base case,
IFGAS_StreamImp is removed and things that inherit from IGAS_StreamImp
are converted to inherit from CFGAS_Stream.

The stream type inputs are changed to IFX_SeekableStream so that they
can accept the same type (IFX_SeekableStream is an
IFX_SeekableWriteStream and an IFX_SeekableReadStream). This way the
storage can be shared in the base class.

Change-Id: I06645071e68e2a4d4120c0e336529f2c18c2b705