Do not walk off end of formcalc string
The fm2js code takes a pointer to the input string and then walks along
that pointer. There are currently no checks to verify we haven't walked
off the end of the pointer into random memory.

If this happens, we can end up allocating large chunks of memory and
copying random bits.

BUG=chromium:721533

Change-Id: Ia61fe96c1ff9eb9ded63cf8326b7be44986bd9e1