Make a raw_ptr type that bothers ASAN
Example.  Not ready for landing, since we need a #define
to turn this off for non-asan builds. Maybe call it
CFX_UnownedPtr.

Half-baked idea from a Friday Afternoon.

There are places where an object "child" has a raw pointer
back to object "owner" with the understanding that owner will
always outlive child.

Violating this constraint can lead to use after free, but this
requires finding two paths: one that frees the objects in the
wrong order, and one that uses the object after the free. The
purpose of this patch is to detect the constraint violation
even when the second path is not hit.

We create a raw_ptr template that is used in place of TYPE*.
raw_ptr's dtor simply goes out and probes the first byte of
the object to which it points.  Used in "child", this allows
ASAN to prove that the "owner" is still alive at the time the
child is destroyed, and hence the constraint is never violated.

Change-Id: I2a6d696d51dda4a79ee2f00a6752965e058a6417