Break linkage from CPDFXA_DocEnvironment to IJS_EventContext.
CPDFXFA_DocEnvionment is part of the heirarchy of objects owned by
the embedder under its FPDF_Document object.

IJS_EventContext is part of the hierarchy of objects owned by the
embedder under its FPDF_FormHandle object.

The FPDF_Document must outlive the FPDF_FormHandle, so there is a
lifetime inversion with the m_pJSEventContext member.  No matter,
event contexts are transient and allocated in a stack-based manner
as control bounces from C++ to JS and vice versa, and keeping a
persistent one seems wrong.  So make it only as needed.

Bug: 723644
Change-Id: I74dd328ad4b501e4adf57821b44ecc78d945eb0f