Bad cast to CPDF_Document in FPDFEditEmbeddertest
The test, as written, is only correct for the non-xfa case, where
the FPDF_Document opaque type is equivalent to the CPDF_Document.
Under XFA, however, there is another layer of indirection that
the test harness needs to respect.

Sorry, I missed this the first time around when I signed off
on the CL that introduced this test.

Change-Id: Ida3579edf639fbfca2939dc151c0420247e90ab3