Make CPDF_ImageCacheEntry retain CPDF_Image
CPDF_PageRenderCache::m_ImageCache is a map from streams
to an image cache entry containing a clump of data associated
with the stream. Oddly, the clump includes the stream key which
we already have in order to get to the clump, but worse doesn't
ensure the existence of the CPDF_Image object which (maybe) owns
the stream key in question.

So replace the stream with a retained ptr to the image.

Bug: 724460
Change-Id: Id13d2c246918d4ff78c12b5bdb927f99c3f5e4e1