Fix CPDF_PageRenderContext cleanup
Make CFX_RenderDevice, which owns the IFX_RenderDeviceDriver, responsible
for restoring driver state as part of its destruction so that its callers
don't have to do so out of turn.

Then re-order CPDF_PageRenderContext destruction order so that
the progressive renderer is destroyed before the device, and the
device destroyed before the options because of unowned pointers
to objects owned by these.

Bug: 726755
Change-Id: I9a6f23da12140b2758b86e6f33f715ad1c679c3f