Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> (4 comments)
> 
> Sorry if I missed it and I'm asking the same question again - what effect does this have on existing callers of FPDFPage_GenerateContent() ?

If a caller is expecting GenerateContent to remove the existing content, then this change could break them. That was why I originally added FPDF_PreserveExistingContent in PS2. It made this change opt-in for the document. That was later removed. I think it's better to just fix it without the opt-in. If someone complains then it should be simple enough to add a public API to explicitly clear the content of a page.

Also, I have seen others mention about this issue as well: 

> "...GenerateContent() will empty a page of everything except images since  all other objects are not added to the buffer" 
> https://groups.google.com/forum/#!topic/pdfium/J1r0IGovfLU

And (in a downstream project):

> "Do not use Generatecontent() method to prevent the loss of page content."
> https://forum.patagames.com/posts/t225-How-to-insert-image-at-the-mouse-click-position-in-WPF-Pdf-Viewer

I think the opt-in would have just made it even more confusing for others.

Patch-set: 12
Reviewer: Ryan Wiley <6990@fca2d64a-0420-3532-b200-12848c95a47b>