Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Ok.  Note that there's no reason for CFGAS_FormatString to exist, it has but one member that points back to the locale.  Alternatively, its ctor might be made to take a CFX_*String* and avoid passing it in on every parsemumble method.


Will take a look in a followup how we use the CFGAS_FormatString, it's possible we could fold the pattern into the constructor or to move everything to static. I'd prefer if it went away, as it isn't just FormatString, it's format and parse.

Patch-set: 3
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>