Update patch set 14
Patch Set 14:

(3 comments)

> Patch Set 14:
> 
> (4 comments)
> 
> Thank you for the CL! I think that there are still important object duplication problems, but this is better than what we had before. I have some late comments, I can do them myself if you don't want to followup on them.

Yes, there is still a duplication issue if a page object in the existing content is modified. Any thoughts on how to solve that? I was thinking we would need to either remove the page object from the old stream or edit it in the old stream and re-write it. Either way that means we would need to keep track of which stream each page object came from along with an offset.

Patch-set: 14