Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I wonder if the test is flaky? There's full coverage here and the win_xfa_clang try bot was fine.
> 
> There's definitely flakiness involved, but I can't tell if there's something about my CL triggering that flakiness or if was just bad luck.

Henrique and I looked at my test and his code together, and we are not sure why this is being flaky. The plan is to hold off this CL for a bit and see if this bot fails for other CL too. Lei, let us know if you have a better suggestion.

Patch-set: 2
CC: Jane Liu <6975@fca2d64a-0420-3532-b200-12848c95a47b>