Added EmbedderTest::RenderPageWithFlags()
Currently, in embedder tests, pages can only be rendered with no flags
specified. With the addition of annotation APIs and corresponding
embedder tests, it is helpful to add a test method to allow flags to be
passed in when rendering pages in embedder tests.

This CL adds such a method.

Bug=pdfium:737

Change-Id: I277831cb2e228154e9a65b078a241b8e61220664