Add FORM_GetSelectedText() function.
This function copies the selected text from a form text field or
form combobox text field into the buffer parameter and returns the
length of the selected text string. When passed buffer is a nullptr
or buflen is less than the length of the selected text, this function
does not modify the buffer and only returns the selected text length.

BUG=chromium:59266

Change-Id: Ie77de38e45bbe6f9ea033826c961435304eedfc7